If-Else.py
https://www.hackerrank.com/challenges/py-if-else/problem?isFullScreen=true

Given an integer, n, perform the following conditional actions:

If n is odd, print Weird
If n is even and in the inclusive range of 2 to 5, print Not Weird
If  is even and in the inclusive range of 6 to 20, print Weird
If  is even and greater than 20, print Not Weird

Input Format

A single line containing a positive integer, n.

Constraints
1<= n <=100
Output Format

Print Weird if the number is weird. Otherwise, print Not Weird.